feat: add authroute and rename pages
