+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
-import { Departures, Landing, Rent, RentCheckout } from './pages'
+import { Arrivals, Checkout, Departures, Landing } from './pages'
 
 const App = () => {
     useEffect(() => {
@@ -13,13 +13,13 @@ const App = () => {
     return (
         <BrowserRouter>
             <Routes>
+                <Route path="/arrivals" element={<Arrivals />} />
+                <Route
+                    path="/checkout/:chainName/:address/:tokenId"
+                    element={<Checkout />}
+                />
                 <Route path="/departures" element={<Departures />} />
                 <Route path="/" element={<Landing />} />
-                <Route path="/rent" element={<Rent />} />
-                <Route
-                    path="/rent-checkout/:chainName/:address/:tokenId"
-                    element={<RentCheckout />}
-                />
             </Routes>
         </BrowserRouter>
     )
