+++ client/src/components/Header.jsx
@@ -24,7 +24,7 @@ export default function Header() {
                 </div>
                 <Popover.Group as="nav" className="hidden md:flex space-x-10">
                     <a
-                        href="/rent"
+                        href="/arrivals"
                         className="text-base font-medium text-gray-500 hover:text-gray-900"
                     >
                         Arrivals
