+++ client/src/pages/index.js
@@ -1,4 +1,4 @@
+export { default as Arrivals } from './Arrivals'
+export { default as Checkout } from './Checkout'
 export { default as Departures } from './Departures'
 export { default as Landing } from './Landing'
-export { default as Rent } from './Rent'
-export { default as RentCheckout } from './RentCheckout'
